+++ lib/clixon/clixon_xml_nsctx.h
@@ -48,12 +48,12 @@
 /*
  * Prototypes
  */
+cvec   *xml_nsctx_init(char *prefix, char *namespace);
+int     xml_nsctx_free(cvec *nsc);
 char   *xml_nsctx_get(cvec *nsc, char *prefix);
 int     xml_nsctx_get_prefix(cvec *cvv,	char *namespace, char **prefix);
-int     xml_nsctx_set(cvec *nsc, char *prefix, char *namespace);
-cvec   *xml_nsctx_init(char *prefix, char *namespace);
+int     xml_nsctx_add(cvec *nsc, char *prefix, char *namespace);
 int     xml_nsctx_node(cxobj *x, cvec **ncp);
 int     xml_nsctx_yang(yang_stmt *yn, cvec **ncp);
-int     xml_nsctx_free(cvec *nsc);
 
 #endif /* _CLIXON_XML_NSCTX_H */